Added email response system for new users.
For this to work you must include the "langKey" parameter with a "language Key" symbol such as "es" => Spanish. 
The string would look like this:
"...": "...",
"langKey": "es".

The email, for the moment, takes a long time to be sent. Obviously you have to enter an existing email so that this can be checked.